WIP - incomplete - moving test out of src. BUT IT'S NOT NECESSARY, so undoing next. Just wanted to keep this state too.
