+++ build-tools/packages/build-tools/src/type-test-generator/main.ts
@@ -17,7 +17,7 @@ import { getFullTypeName, getNodeTypeData, TypeData } from "../typeValidator/typ
 // Do not catch error from opening file since the default behavior is fine (exits process with error showing useful message)
 const packageObject: PackageJson = readJsonSync("package.json");
 
-const testPath = `./src/test/types`;
+const testPath = `./test/types`;
 // remove scope if it exists
 const unscopedName = path.basename(packageObject.name);
 
@@ -112,7 +112,7 @@ const testString: string[] = [
  * Generated by fluid-type-test-generator in @fluidframework/build-tools.
  */
 import * as old from "${previousPackageName}";
-import * as current from "../../index";
+import * as current from "../../src/index";
 
 type TypeOnly<T> = {
     [P in keyof T]: TypeOnly<T[P]>;
@@ -170,7 +170,7 @@ for (const oldTypeData of previousData) {
 	testString.push("");
 }
 
-mkdirSync("./src/test/types", { recursive: true });
+mkdirSync("./test/types", { recursive: true });
 
 writeFileSync(filePath, testString.join("\n"));
 console.log(`generated ${path.resolve(filePath)}`);
