+++ packages/dds/sequence/.eslintrc.js
@@ -6,7 +6,7 @@
 module.exports = {
 	extends: [require.resolve("@fluidframework/eslint-config-fluid/minimal"), "prettier"],
 	parserOptions: {
-		project: ["./tsconfig.json", "./src/test/tsconfig.json"],
+		project: ["./tsconfig.json", "./test/tsconfig.json"],
 	},
 	rules: {
 		"@typescript-eslint/no-use-before-define": "off",
