+++ packages/dds/sequence/test/intervalCollection.fuzzUtils.ts
@@ -10,8 +10,8 @@ import {
 } from "@fluid-internal/stochastic-test-utils";
 import { DDSFuzzTestState } from "@fluid-internal/test-dds-utils";
 import { PropertySet } from "@fluidframework/merge-tree";
-import { IntervalType } from "../intervalCollection";
-import { SharedStringFactory } from "../sequenceFactory";
+import { IntervalType } from "../src/intervalCollection";
+import { SharedStringFactory } from "../src/sequenceFactory";
 
 export interface RangeSpec {
 	start: number;
