+++ packages/dds/sequence/test/intervalCollection.snapshot.spec.ts
@@ -10,14 +10,14 @@ import {
 	MockStorage,
 } from "@fluidframework/test-runtime-utils";
 import { ISummaryTree } from "@fluidframework/protocol-definitions";
-import { SharedString } from "../sharedString";
-import { SharedStringFactory } from "../sequenceFactory";
+import { SharedString } from "../src/sharedString";
+import { SharedStringFactory } from "../src/sequenceFactory";
 import {
 	IntervalCollection,
 	intervalLocatorFromEndpoint,
 	IntervalType,
 	SequenceInterval,
-} from "../intervalCollection";
+} from "../src/intervalCollection";
 
 const assertIntervals = (
 	sharedString: SharedString,
