+++ packages/dds/sequence/test/intervalCollection.spec.ts
@@ -14,9 +14,9 @@ import {
 	MockStorage,
 	MockEmptyDeltaConnection,
 } from "@fluidframework/test-runtime-utils";
-import { SharedString } from "../sharedString";
-import { SharedStringFactory } from "../sequenceFactory";
-import { IntervalCollection, IntervalType, SequenceInterval } from "../intervalCollection";
+import { SharedString } from "../src/sharedString";
+import { SharedStringFactory } from "../src/sequenceFactory";
+import { IntervalCollection, IntervalType, SequenceInterval } from "../src/intervalCollection";
 
 const assertIntervals = (
 	sharedString: SharedString,
