+++ packages/dds/sequence/test/marshalling.spec.ts
@@ -5,7 +5,7 @@
 
 import { strict as assert } from "assert";
 import { Marker, ReferenceType, TextSegment } from "@fluidframework/merge-tree";
-import { SubSequence } from "../sharedSequence";
+import { SubSequence } from "../src/sharedSequence";
 
 const segmentTypes = [
 	{
