+++ packages/dds/sequence/test/memory/sharedSequence.spec.ts
@@ -4,7 +4,7 @@
  */
 
 import { benchmarkMemory, IMemoryTestObject } from "@fluid-tools/benchmark";
-import { SubSequence } from "../../sharedSequence";
+import { SubSequence } from "../../src/sharedSequence";
 
 describe("SharedSequence memory usage", () => {
 	// IMPORTANT: variables scoped to the test suite are a big problem for memory-profiling tests
