+++ packages/dds/sequence/test/snapshots/emptyPropsAtEnd.json
