+++ packages/dds/sequence/test/snapshots/legacy/headerAndBody.json
