+++ packages/dds/sequence/test/snapshots/legacy/headerOnly.json
