+++ packages/dds/sequence/test/snapshots/legacy/withAnnotations.json
