+++ packages/dds/sequence/test/snapshots/legacy/withIntervals.json
