+++ packages/dds/sequence/test/snapshots/legacyWithCatchUp/headerAndBody.json
