+++ packages/dds/sequence/test/snapshots/legacyWithCatchUp/headerOnly.json
