+++ packages/dds/sequence/test/snapshots/legacyWithCatchUp/withAnnotations.json
