+++ packages/dds/sequence/test/snapshots/legacyWithCatchUp/withIntervals.json
