+++ packages/dds/sequence/test/snapshots/v1/headerAndBody.json
