+++ packages/dds/sequence/test/snapshots/v1/headerOnly.json
