+++ packages/dds/sequence/test/snapshots/v1/largeBody.json
