+++ packages/dds/sequence/test/snapshots/v1/withAnnotations.json
