+++ packages/dds/sequence/test/snapshots/v1/withIntervals.json
