+++ packages/dds/sequence/test/snapshots/v1Intervals/withV1Intervals.json
