+++ packages/dds/sequence/test/subSequence.spec.ts
@@ -12,7 +12,7 @@ import {
 } from "@fluidframework/merge-tree";
 // eslint-disable-next-line import/no-internal-modules
 import { TestClient } from "@fluidframework/merge-tree/dist/test";
-import { SubSequence } from "../sharedSequence";
+import { SubSequence } from "../src/sharedSequence";
 
 const clientNames = ["Ed", "Ted", "Ned", "Harv", "Marv", "Glenda", "Susan"];
 const verbose = true;
