+++ packages/dds/sequence/test/tsconfig.json
@@ -1,8 +1,8 @@
 {
 	"extends": "@fluidframework/build-common/ts-common-config.json",
 	"compilerOptions": {
-		"rootDir": "./",
-		"outDir": "../../dist/test",
+		"rootDir": ".",
+		"outDir": "../dist/test",
 		"types": ["mocha"],
 		"declaration": false,
 		"declarationMap": false,
@@ -12,7 +12,7 @@
 	"include": ["./**/*"],
 	"references": [
 		{
-			"path": "../..",
+			"path": "..",
 		},
 	],
 }
