+++ packages/dds/sequence/tsconfig.json
@@ -1,6 +1,5 @@
 {
 	"extends": "@fluidframework/build-common/ts-common-config.json",
-	"exclude": ["src/test/**/*"],
 	"compilerOptions": {
 		"rootDir": "./src",
 		"outDir": "./dist",
